chore: Rename `model` to `api_model`
This model is *only* for types that are exposed on the API.

We should wrap the API classes properly, if we expose internal things the API structs / enums can get unnecessarily complicated/complex and we might run into trouble when using them in Dart.
In order to make it clear what this "model" is about the module should have a specific name - we should not use any of the types in `api-_module` in the business logic layer (`trade` module, `event` module, ...).
