+++ mobile/native/src/lib.rs
@@ -1,8 +1,8 @@
 mod api;
+pub mod api_model;
 mod bridge_generated;
 pub mod calculations;
 pub mod event;
 pub mod ln_dlc;
 pub mod logger;
-pub mod model;
 pub mod trade;
